Casts node display state to integer in node_dev
Adds an explicit `static_cast` when updating the device state from the node's display state. This ensures the value matches the expected parameter type for `direct_set_state` and avoids implicit conversion warnings.

+++ PAC/common/device/node_dev.cpp
@@ -20,7 +20,8 @@ void node_dev::evaluate_io()
     if ( node )
         {
         // Обновление состояния на основе состояния узла.
-        device::direct_set_state( node->get_display_state() );
+        device::direct_set_state( 
+            static_cast<int>( node->get_display_state() ) );
         }
     else
         {
